Migrate OAuth from Rust to frontend (#1)
Let's aim to have a really minimal Rust backend, and handle most of the logic in the frontend. This way, we don't have to sync types between Rust and TS, and we can take advantage of great TS Oauth libraries like arcticjs and zod for validation.

* Move routes

* Move to providers

* Move to providers.google

* Use frontend for google oauth

Keep the rust side very minimal

* Move refreshToken to frontend

* Move endpoint for fetching calendars

* Move sync events to ts too